requirements.txt
asgiref==3.4.0
certifi==2022.6.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
dill==0.3.5.1
Django==3.2.15
django-environ==0.9.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.1.0
drf-yasg==1.21.3
flake8==5.0.4
idna==3.3
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
packaging==21.3
platformdirs==2.5.2
psycopg2==2.9.3
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2021.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tomlkit==0.11.4
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.12
wrapt==1.14.1
